move frozen to rustc_data_structures
